Rename VertexArrayType enum entries for clarity
The VertexArrayType entries loosely referred to the kind of data for each attribute,
but this was inconsequential.  Rename them so that they refer to the types of the
attributes rather than the kinds.  For example V3N3C3 becomes F3F3F3.  The code
generating the attribute data and the GLSL consuming it determines its actual use.